Remove Company v1 views
They haven't been used by the front end for a bit, so it's safe to get rid of them now.
