+++ datahub/metadata/serializers.py
@@ -5,14 +5,6 @@ from datahub.core.serializers import NestedRelatedField
 from .models import Country, TeamRole, UKRegion
 
 
-class NestedCountrySerializer(serializers.ModelSerializer):
-    """Nested Country serializer."""
-
-    class Meta:  # noqa: D101
-        model = Country
-        fields = '__all__'
-
-
 class TeamSerializer(serializers.Serializer):
     """Team serializer."""
